changed files in server

+++ server/src/controllers/jobApplicationController.ts
@@ -8,10 +8,8 @@ import { StatusCodes } from "http-status-codes";
 
 const createJobApplication = asyncHandler(
   async (req: Request, res: Response) => {
-    // Validate job application data
     const result = validateJobApplication(req.body);
 
-    // If validation fails, throw a custom error with detailed message
     if (!result.success) {
       const errorMessage = result.error.issues
         .map((issue) => issue.message)
@@ -19,12 +17,10 @@ const createJobApplication = asyncHandler(
       throw new CustomError(errorMessage, StatusCodes.BAD_REQUEST);
     }
 
-    // Destructuring validated data
     const { company, role, status, appliedDate, link } =
       req.body as IJobApplication;
 
     try {
-      // Create the job application document
       const jobApplication = await JobApplication.create({
         company,
         role,
@@ -33,10 +29,8 @@ const createJobApplication = asyncHandler(
         link,
       });
 
-      // Respond with success
       res.status(StatusCodes.CREATED).json(jobApplication);
     } catch (error) {
-      // Handle unexpected errors (e.g., database issues)
       console.error("Error creating job application:", error);
       throw new CustomError(
         "Failed to create job application",
@@ -72,19 +66,24 @@ const deleteJobApplication = asyncHandler(
     const { id } = req.params;
 
     try {
-      // Find the job application by ID and delete it
+      if (!id) {
+        res.status(StatusCodes.BAD_REQUEST).json({
+          message: "Job application ID is required",
+        });
+        return;
+      }
+
       const jobApplication = await JobApplication.findByIdAndDelete(id);
 
-      // If job application doesn't exist, return 404
       if (!jobApplication) {
         res
           .status(StatusCodes.NOT_FOUND)
           .json({ message: "Job application not found" });
-      } else {
-        res
-          .status(StatusCodes.OK)
-          .json({ message: "Job application deleted successfully" });
+        return;
       }
+      res
+        .status(StatusCodes.OK)
+        .json({ message: "Job application deleted successfully" });
     } catch (error) {
       console.error("Error deleting job application:", error);
       res
@@ -96,26 +95,34 @@ const deleteJobApplication = asyncHandler(
 
 const updateJobApplication = asyncHandler(
   async (req: Request, res: Response) => {
+    const result = validateJobApplication(req.body);
+
+    if (!result.success) {
+      const errorMessage = result.error.issues
+        .map((issue) => issue.message)
+        .join(", ");
+      throw new CustomError(errorMessage, StatusCodes.BAD_REQUEST);
+    }
+
     const { id } = req.params;
+    
     const { company, role, status, appliedDate, link } =
       req.body as IJobApplication;
 
     try {
-      // Find the job application by ID and update it with new values
       const updatedJobApplication = await JobApplication.findByIdAndUpdate(
         id,
         { company, role, status, appliedDate, link },
-        { new: true } // Return the updated document after the update operation
+        { new: true }
       );
 
-      // If job application doesn't exist, return 404
       if (!updatedJobApplication) {
         res
           .status(StatusCodes.NOT_FOUND)
           .json({ message: "Job application not found" });
-      } else {
-        res.status(StatusCodes.OK).json(updatedJobApplication);
+        return;
       }
+      res.status(StatusCodes.OK).json(updatedJobApplication);
     } catch (error) {
       console.error("Error updating job application:", error);
       res
@@ -125,4 +132,9 @@ const updateJobApplication = asyncHandler(
   }
 );
 
-export { createJobApplication, getAllJobApplications, deleteJobApplication, updateJobApplication };
+export {
+  createJobApplication,
+  getAllJobApplications,
+  deleteJobApplication,
+  updateJobApplication,
+};
